requirements.txt
discord.py==1.2.5
discord.py[voice]
google-api-python-client
google-auth
google-auth-httplib2
oauth2client
pynacl
youtube_dl
asyncio
praw
aiohttp
